5 задание, разбил код на модули
